added Prompt class, fixed some small structural issues
The Prompt class is just a line starting with a label and ending with a
value. It is selectable, like ListView.

There is now also proper `__repr__` code for most elements. It provides
a string that can be eval-d (please don't) to create said object, as far
as constructor arguments go anyways.