moved all widget code to a submodule
widgets.py got over 1000 lines, and that was no bueno. Now there is a
submodule that contains 2 sub-submodules?

There is base.py, which contains all the base widgets, and extra.py,
which provides some extra things that are created using the widgets
mentioned above.

Of course, all of them rely on the Widget class, but that doesn't matter
:).