better windows support
getch() now returns proper string values, and can be used with other functions, like strip_ansi() or real_length()

This is done with a bit of a hack, as msvcrt.getch() returns bytes. The first character is not a unicode-decodable, so we substitute it for the escape character used in UNIX systems.

Color support is not great, since cmd nor powershell 7 support ANSI 8-bit colors natively. RGB colors work with a hack, but it's not implemented as it would break some other things. We should come back to this.