added a Splitter object
This class allows you to display elements side-by-side. It's pretty
neat, but has some flaws for now.

Some smaller additions include renaming `Widget().dbg()` to
`Widget().debug()`, and some minor changes in how Containers add
widgets.

For one, two objects with differing heights will cause the Splitter to
only show as many lines as there are in the shortest element.

The other big issue is that ListView doesn't work properly with this
class.

As a personal side-note these past two days of work on this module were
a really nice boost for my confidence in myself. The biggest issues
with the previous iteration of this library were the pretty bad
implementation of the InputField object, and it not allowing multiline
inputs, and the constraint of only being able to stack objects
vertically.

In the past 2 days, both of these issues are (more or less). It might
not look like much, but they were the kind of things that I fully gave
up on implementing previously, and now they are fully working in less
than 48 hours.

Good stuff.