improved the InputField API
Added new methods `has_selection()` to query if there is a selected
range set for the object, and `select()` to allow a nicer interface for
setting selected ranges.

There is also nicer call for range selections. Now you can call
`select_range((start, end))` to start a selection, and
`clear_selection()` to reset this setting.

The second method also required adding an argument type of `tuple[int,int]`
to the base Widget.

Along with these changes is some minor fixes to how newlines are drawn,
as well as making `highlight_style` optional, with the object defaulting
to cursor style if it is set to `None`.