added `prettify_markup()` and improved tokenization
Tokens now have an additional field, `attribute`, that is chosen from
the `TokenAttribute` enum. It signifies what special meaning the token
has, which helps with syntax highlighting it.

For example, tokens starting with `/` are given
`TokenAttribute.CLEAR`, showing syntax highlighters their special value.

The `prettify_markup()` function is essentially a syntax highlighter for
markup (how unexpected, I know.). Currently it is used for the command
line interface, but it can probably be of use in other places too.