added the Serializer() class! :)
This class is used to, well, serialize objects. It is a whole class
instead of just some functions, because it needs to be able to be aware
of custom objects, not just base pytermgui ones.

Originally it would've provided methods named after json's, since they
are pretty universally known for the Python community. I decided against
doing so however, because the whole library is made to be really verbose
and readable, and 'dumps' is just not that. :D

The serializer internally calls `Widget().serialize()`, which in turn goes
through `Widget()._serialized_fields`, which is by default a copy of the
class' `Widget.serialized` attribute.

You can specify style values by adding them as '*style_name'. This will
be serialized as `Widget().get_style('style_name')('depth', 'value')`.
This allows baking styles into an object, though the actual style
callable won't be saved, only the styled value. These values also go
through `markup_to_ansi()`, so the values are nicely readable, though
not too optimized (coming soon!).