created the `_IDManager()` class!
This class is used to store and access all id-registered widgets.

On load of the widgets submodule, an instance of `_IDManager` is
created, and assigned to the name `manager`, so you can import and use
it later. The `Widget` class gets this object assigned to it as its
`_manager`, so all widgets use the same object.

Whenever a widget has its `id` property set, it 'phones home' to the
`manager` object, and registers itself. Once that is done,
you can use the `manager` object to `get_widget()` if you know its id,
or `get_id()` if you don't.

This will be useful for layout files, and probably other applications
too.