actually added the serializer!
So I was intelligent, and didn't `git add` the serializer file earlier,
and noticed my mistake when the file wasn't available to me anymore.

To fix this, I did the only logical thing, and rewrote the entire file
again. Props of this is that much of the logic is even more refined,
which is always cool.

One small change is that the class has been renamed to `_Serializer`,
and the only thing that's exported is an already bound object, with the
name `serializer`. This just makes it work like the `_IDManager` or
color objects.