simplified the Widget-internal styling interface
Previously you'd have to call a style method with self.depth and the
item you want it to apply to, which led to a lot of repeated code. Now
`get_style` returns an internal method that only takes `item` as its
argument, and automatically calls the actual method with the right
depth.

This creates two small problems:
    - depths can *technically* change between start of `get_lines` and
      application, however since it's a couple milliseconds only this
      should not really happen. (especially in a single-threaded
      environment, which I expect to be the majority of ptg code)

    - the original method now gets lost when trying to inspect the
      `get_style` return. The only way to fix this without implementing
      an entire class for it was to set the `origin` attribute of the
      `_apply_style` method to the originally supplied style. This at
      least allows the syntax of `obj.get_style("value").origin`.