`markup_style` all the things!
All instances of `default_foreground` have been replaced by
`markup_style`, and some minor shuffling-arounds have been made to
accomodate this change.

The minor side-effect of this is that since this is applied to all
characters, every border char of Container has the same repeated codes.
This will be fixed in an upcoming commit.