Added better mouse input API
There is a new context manager, `mouse_handler`, which takes an event to
listen on, and yields `translate_mouse`. The use of this method is to
more easily and cleanly handle mouse events, as this stops the reporting
on exit.

Also added an implicit onclick method to all `Widget`-s, which allows
them to be selected through the root `Container` and focused, without
the user having to specify an onclick method for it.

The last big issue is that this system does NOT support nested elements,
such as multiple Containers within eachother or even just a Splitter
having clickable elements. That is a TODO.