Added `!align` markup macro!
This is essentially an interface to the f-string alignment methods. It
supports aligning left, right & center, using the syntax:

"<width>:<alignment> <content>" -> "[!align]30:center This is centered"

Note: Content cannot have inner markup, as it breaks a lot of things.
This is to be fixed some time in the future.