Improved the mouse translation methods
There is a new Enum, `MouseAction`, which at this time is one of
`PRESS`, `RELEASE`, `HOLD, `SCROLL_UP`, `SCROLL_DOWN`. It is return
alongside the position by the `translate_mouse` method.

Also, there is a cool window-manager-like demo in tests/mouse_test.py.
You can drag the top bar of the window to move it. In the future, I'm
planning on making a full demo including support for unlimited windows
and changing their properties in the window itself.