A new `get_lines()` method for `Container`!
This is the backbone of the entire widget-system, so it is important to
keep it readable & expandable, which it failed at doing thus far.

Some things have changed in this rewrite, however. The only user-side
effect is that instead of `Container` global horizontal alignment, the
`Widget` class has its own `parent_align` value, which can be set as a
default value using `Widget.DEFAULT_PARENT_ALIGN`.

Another big change is that vertical alignment within `Container`-s is
now gone, at least in its old form. This had to be done because updating
the widget position was rather impossible in this setup, as the
alignment has to be done after the widget lines are acquired, and the
aligned lines have no link to the original `Widget`-s.

This is to be revisited in the future, the current idea is that there
will be a `Container.center_elements()` or similar method, which will
insert padding labels into the element. This also might see the
introduction of a simple `Padding` class, but that is to be seen.

All in all, this is a really nice change, and will help us a lot in the
future, and allow for more possibilities.

ok, bye.