Added `magic` type conversion to `Container._add_widget`
You can now add non-widgets to a `Container`, and as long as they are a
convertable type (see `help(pytermgui.auto)`) a new Widget will be
created in their place.

This for now means you can add `str`-s in place of `Label`-s,
`list[str]`-s in place of `ListView`-s and `dict[str, str]` in place of
`Prompt` (though the latter class does not work at the moment).