Improved `Button`
The styling is a lot more consistent now, though it will suffer the loss
of delimiter styling. This was needed because doing any markup on
delimiter char `[` forced it to be escaped, and because of how things
worked that part had to happen for it all to work.

This may be revisited in the future though.