Refactored `WindowManager.run()`, added safe state fallback to Windows
The previous implementation used 3 (!) pylint error silencers, which
really says it all. It was doing a lot more than it should've, and now
that is fixed.

The function's inner body was split into 2 new functions, with the
original only calling these and exiting when necessary. Similar to
`Window`, there are now `handle_key` and `handle_mouse` methods, the
former mainly handling bindings, and the latter implementing window
movement & resizing.

As a new feature, every time a window is resized, it is now checked for
its state (`width`, `forced_width`) being valid. If it is, the window
manager caches this state inside the window object, and whenever it gets
into a state where `Window.get_lines()` fails it restores it. This is
ultimately a bandaid on the larger issue of mouse inputs happening
quicker than `Window`-s can resize.

ok, bye.