Add `Widget.bind` method
This method already existed for `WindowManager` and `InputField`, so it
felt nicer to have it as a built-in function. Binding works on an opt-in
policy by each widget, it raises a `TypeError` if a widget's
`_is_bindable` flag is not set to True.

The signature of this callback has also changed. It is now universally
`Callable[[Widget, str], Any]`, meaning it gets the widget whos binding
we activated, and the key that activated said binding. Return values are
currently disregarded.