Rewrite `cmd.py` to use an `Application` framework
With this, calling `ptg` from the command line drops you into a
`WindowManager` with all of the utility-apps the library provides. For
now, these are `getch` and `markup`.

These apps can also be launched in standalone mode (`ptg --app {getch,
markup}`), which causes the WM to exit after the window is done with its
input (pressing return in markup, pressing any key in getch), and dump
its output.

The framework is based on the `Application` class. Each app subclasses
said application, providing a `construct_window()` and a `finish()`
method. The first one is called whenever the app is launched, while the
latter is called after WM exit in standalone mode.

This structure will likely be used in `teahaz` some point down the line.