Implement new `window_manager.py` module
This module originated as a fun weekend project, so it lacked in code
quality, performance and general structure. This rewrite aims to solve
all of that.

The biggest update is in performance: The rendering is now frame-based,
and is not greedy (doesn't wastefully render every cycle it gets). There
is also various optimizations included, such as not printing when no
changes have occured, and caching non-focused window display data.