Move mouse-handling task from `WindowManager` -> `Widget`
Previously, all mouse handling went through the `WM`, and no `Widget`-s
could do anything special. With this change, `Widget`-s get mouse events
first, and are therefore able to add custom behaviours.

If a `Widget`'s `handle_mouse` method returns `True`, the event is
assumed to be handled and `WindowManager` never sees it. If it returns
`False`, the event is passed through to `WindowManager`, AND the widget
gets selected. In some cases the latter behaviour is not desired, so it
can be avoided by simply returning `False` in the handler.