Fix `file_loader` id support
This change required changing the `Namespace.__getattr__()`
functionality. Previously, it would return a copy of the `Widget` it
found rather than the original, but it no longer does so. This is
because the id system didn't seem to want to work around this idea, and
at the end of the day it was kind of useless to begin with. Usually, the
user wouldn't mind the original `Widget`, and if they need to have
multiple they can just `copy()` it.

This change will be reflected in the documentation soon.