Improve how WindowManager resizing is handled
Previously, resizing the terminal would force all Windows (and
Containers) to become centered, even if they weren't supposed to be.
This is no longer the case, and terminal resizing is now fully handled
by WindowManager.

For now, whether a window was static or not is not really checked when
moving them. This is probably the better behaviour, as if a static
window goes off-screen printing will get fully messed up as a result.
May revisit this in the future.