Refactor compositing & drawing routine for more flexibility
The FPS target still isn't hit very accurately, usually under-pinning by
about 10%, though the difference drops as the target increases.

We can either fix the root cause, or come up with a function that more
or less models this difference over framerate. Not the right solution,
but definitely the more fun one.