Fix `terminal.size` being offset by origin
The reason this existed in the first place is my assumption that the
origin of the terminal isn't factored into the size query, when it is.

Instead of returning a size with the origin removed, we apply the origins
when clamping window positions. This way the program can actually take
up the full space.