Improve scrolling performance
There used to be a need to re-run each widget's `get_lines` method
after positioning them, but it seems to no longer be an issue. The
ideal end-goal IMO would be a version of the `get_lines` method that
stops when it goes outside of the range visible in the scroll area.