Add `highlighters` module
This can be used to highlight Python code at the moment, but adding
other languages shouldn't be so hard. It uses regex under the hood,
and it's surprisingly fast and efficient as a result.

This is NOT a full syntax-highlighter, and there are no plans to make
it one. It lacks contextual awareness, so multiline comments do not
work at the moment. Complex (containing escaped quotes) strings also
mess it up, but that's more of a problem with not finding the right
pattern yet, and probably will eventually be fixed.