Implement blurred styles, remove `Window.allow_fullscreen`, add `Collapsible`
The new blurred styles take the place of the previous, kind of buggy
and pretty slow full-window markup strip effect. It's not quite perfect
yet, but the basic framework is there. There is also a wrapper API to
set these styles easier, but the `styles.` syntax still works.

`Window.allow_fullscreen` was always a bit of an odd item. It used to
be applied by the window manager, which was a bit odd but acceptable,
but since the rewrite it got handed off to the compositor, which should
have no right to mess with any settings. The new layout system completely
removed the need to have this attribute, and underlying mess, so it's
gone.

The `Collapsible` widget is really just a way to have things "hidden"
and revealed dynamically in the UI. Again, the API isn't perfect yet,
but it's needed in `cmd` and I thought I might as well make it a public
thing at that point. More work will be done on it, however.

:)