Fix issues related to the recent fix to #49
Turns out the problem was only _partially_ about missing a default
width, and it was much more related to the recent change to how
TIM handles escaped markup.

Previously, the `\` symbol remained in the output. This was a bit ugly,
and usually not what you wanted, so it was changed to omit that sign. As
a result, once an escaped markup tag was parsed, the output was no longer
escaped.

The fix was pretty simple, once the issue was clear: Instead of doing
3 parses within the button get_lines, we now parse only a single time.