Cache `MarkupFormatter` `get_markup` results
When a markup formatter is called, it always uses the markup form of
the given ANSI-styled text for stability reasons. However, this
operation can be quite slow, so for the moment we will cache all
input styled text and its returned markup.

This _might_ cause some issue later down the line. It is assumed that
while markup -> ANSI conversions may be volatile (e.g. change
throughout runtime), the reverse conversion _should_ be the same
always. We will see if this is a problem, but regardless a TIM-side
cache would probably be more useful.