Remove default `wm-title` alias from `Window.set_title`
This used to be a defined alias in the days of the old `cmd`, but
no longer is. There is a replacement, `ptg.title`, but that only
gets defined in the `cmd` context, so is useless to everyone else.

If needed/wanted, a style can be manually inserted into the title
string:

  window.set_title("[my_app.title] My Title [/my_app.title]")