Introduce `__fancy_repr__` dunder method
This method can be used to have some pretty-repls without breaking
the normal ones used by most of python. It works similarly to that
of rich, whereas it's based on yielding a set of values that are all
appended to one another.

Each yield can either be a dict in the format:

  {"text": <your_text>, "highlight": <bool>}

or a shorthand of just a string, which is interpreted to mean:

  {"text": <the_str_you_gave>, "highlight": False}

In the future the `highlight` argument could be expanded to be a str,
which would be the name of a highlighter class. Since we only really
have python at the moment this isn't too worth implementing, but at
some point an example `__fancy_repl__` yield could be:

  {"text": <your_text>, "highlight": "python"}