Fix unicode block characters not rendering in the correct spot
SVGs by default don't render characters to the origin of their "box",
but somewhere below. This is usually the behaviour we want, however
block characters start to look a bit silly when they don't actually
fill up the part of the screen they should be.

This is fixed by setting `alignment-baseline: text-after-edge` in styles,
and then applying an offset to the background rects in order to keep
them in the right place, relative to the characters they are behind.