Add `tokenize_ansi` function
This tokenizer is quite a bit more advanced than what came before it;
it can now parse any number of compound ANSI sequences.

For example

  \x1b[38;5;141;1;3m

Stands for purple text, that is bold and italic. This didn't use to be
parseable, as it would only recognize the `38;5;141` part of the sequence.