[manager]: Stop prolonged blocking on `getch` calls (fixes #95)
The solution was to switch to timeout-based calls, which let us fix the
problem on POSIX and not bring up a new problem on Windows. The reason
it works is a bit sketchy; Windows exposes API to detect no available
input before calling `getch`, so instead of applying the timeout it
returns immediately, which is perfect for our cause.